requirements.txt
aiohttp
requests
pycryptodome
python-dotenv
fastapi
uvicorn
curl-cffi
pydantic
pillow
colorama
esprima
